% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_util.R
\name{compute_auc}
\alias{compute_auc}
\title{Calculate AUC for ROC}
\usage{
compute_auc(curve)
}
\arguments{
\item{curve}{output of compute_tpr_and_fpr}
}
\value{
a data.frame with one row and one column named roc_auc
}
\description{
Take the ROC curve output from compute_tpr_and_fpr function and compute the AUC of the curve
}
