% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_roc_pr.R
\name{compute_power_and_fdr}
\alias{compute_power_and_fdr}
\title{Calculate power and FDR}
\usage{
compute_power_and_fdr(discovered_genes, true_genes)
}
\arguments{
\item{discovered_genes}{a set of proposed signals (their IDs as string)}

\item{true_genes}{a set of true signals (their IDs as string)}
}
\value{
data.frame with power and fdr as columns along with nTP (number of true positives), nP (number of positives, i.e. proposed signals), and nT (number of true signals)
}
\description{
Given the set of true signals and the set of proposed signals, calculate power and false discovery rate
}
\examples{
compute_power_and_fdr(
  discovered_genes = c('g1', 'g3', 'g4'),
  true_genes = c('g1', 'g2')
)

}
