% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_roc_pr.R
\name{compute_tpr_and_fpr}
\alias{compute_tpr_and_fpr}
\title{Calculate TPR and FPR}
\usage{
compute_tpr_and_fpr(tested_genes, discovered_genes, true_genes)
}
\arguments{
\item{tested_genes}{a set of candidate genes (their IDs as string; it should include both discovered_genes and true_genes)}

\item{discovered_genes}{a set of proposed signals (their IDs as string)}

\item{true_genes}{a set of true signals (their IDs as string)}
}
\value{
data.frame with tpr and fpr as columns along with nTP (number of true positives), nT (number of true signals), nFP (number of false positives, i.e. proposed but not true signals), and nTP (number of proposed and true signals)
}
\description{
Given the set of all candidate genes, true signals, and proposed signals as three separate sets, calculate true positive rate (TPR) and false positive rate (FPR)
}
\examples{
compute_tpr_and_fpr(
  tested_genes = c('g1', 'g2', 'g3', 'g4', 'g5', 'g6'),
  discovered_genes = c('g1', 'g3', 'g4'),
  true_genes = c('g1', 'g2')
)

}
