% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_roc_pr.R
\name{gen_fdr_power_curve}
\alias{gen_fdr_power_curve}
\title{Generate precision-recall curve}
\usage{
gen_fdr_power_curve(true_genes, gene, score, method = "gt", cutoff = NULL)
}
\arguments{
\item{true_genes}{the set of true signals}

\item{gene}{the set of all candidate signals}

\item{score}{the scores of the candidates (the same length as gene)}

\item{method}{method = 'gt' to propose score greater than threshold; method = 'lt' to propose score less than threshold}

\item{cutoff}{cutoff = NULL will use all unique scores from true signals as cutoffs (essentially this is the finest grid we can get)}
}
\value{
a data.frame with all columns of compute_power_and_fdr's output along with cutoff column telling which cutoff is used to obtain the results of the row and precision (same as 1 - fdr) and recall (same as power)
}
\description{
Given the set of true signals and the set of candidate genes along with their scores and proposing method, generate precision-recall curve, which is a sequence of precision and recall measured at a set of cutoffs
}
\examples{
gen_fdr_power_curve(
  true_genes = c('g1', 'g3', 'g4'),
  gene = c('g1', 'g2', 'g3', 'g4', 'g5', 'g6'),
  score = 1:6,
  method = 'gt',
  cutoff = NULL
)

}
