% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_standard-data.R
\docType{data}
\name{gene_annotation_gencode_v26_hg38}
\alias{gene_annotation_gencode_v26_hg38}
\title{Gene annotation (position and gene name etc)}
\format{a list of :
  1. gene_annotation, data.frame with columns:
    1.1. gene_id for Ensambl ID (no dot)
    1.2. gene_name for the "conventional" name of the gene
    1.3. gene_type for type of gene, see \href{https://www.gencodegenes.org/pages/biotypes.html}{here} for more information
    1.4. chromosome
    1.5. start (base-1, included)
    1.6. end (base-1, included)
    1.7. strand: strand of the transcript (optional)
  2. meta_information, a list of
    2.1 genome_assembly_version
    2.2 notes}
\usage{
data(gene_annotation_gencode_v26_hg38)
}
\description{
Gene annotation obtained from gencode v26 (formatted by Alvaro Barbeira).
It uses hg38 genomic coordinate.
}
\examples{
data(gene_annotation_gencode_v26_hg38)

}
\keyword{datasets}
