% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{greedy_map}
\alias{greedy_map}
\title{Greedy mapping}
\usage{
greedy_map(map_table, silver_table, trait_code)
}
\arguments{
\item{map_table}{a data.frame containing 'trait' as column along with 'phecode' and/or 'HPO' as columns.
This table serves as the map between trait name in 'trait' column and trait codes ('phecode', 'HPO', etc).
So that it is used to map trait names (in score_table) to trait code (in silver_standard).}

\item{silver_table}{a data.frame where each row pairs trait in phecode, HPO to gene.}

\item{trait_code}{trait code to use for mapping: 'phecode' or 'HPO'.}
}
\value{
a data.frame pairing trait name to gene
}
\description{
Map trait name to gene if the trait name shared at least one trait_code with the gene, where
trait name to trait_code is specified in map_table and
gene to trait_code is specified in silver_table
}
