% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlib_roc_pr.R
\name{is_sig}
\alias{is_sig}
\title{Check if the candidates should be proposed}
\usage{
is_sig(score, threshold, method, tie = T)
}
\arguments{
\item{score}{the scores of candidates}

\item{threshold}{threshold to propose a signal}

\item{method}{method = 'gt' to propose score greater than threshold; method = 'lt' to propose score less than threshold}

\item{tie}{if propose the candidates which have scores equal to threshold (default TRUE)}
}
\value{
a binary vector of the same length as the input score vector indicating if the corresponding score entry is proposed
}
\description{
Given the scores of the candidates and a threshold to apply along with method to propose, e.g. 'gt' means propose if score is greater than the threshold and 'lt' means propose if score is less than the threshold. It handle 'tie', when tie = TRUE, the candidate with score equals to threshold will be proposed
}
\examples{
is_sig(score = 1:10, threshold = 3, method = 'gt', tie = TRUE)

}
