% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_standard-data.R
\docType{data}
\name{ld_block_pickrell_eur_b38}
\alias{ld_block_pickrell_eur_b38}
\title{LD block annotation (for European population, in hg38)}
\format{a list of
  1. ld_block, a data.frame with columns (each row is an LD block):
    1.1. region_name name of the LD block
    1.2. chromosome
    1.3. start (base-1, included)
    1.4. end (base-1, included)
  2. meta_information, a list of
    2.1 genome_assembly_version
    2.2 notes}
\usage{
data(ld_block_pickrell_eur_b38)
}
\description{
Obtained from \href{https://academic.oup.com/bioinformatics/article/32/2/283/1743626}{Berisa et al}.
Formatted by Alvaro Barbeira
}
\examples{
data(ld_block_pickrell_eur_b38)

}
\keyword{datasets}
