% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_standard-data.R
\docType{data}
\name{orphanet_based_silver_standard}
\alias{orphanet_based_silver_standard}
\title{Silver standard based on Orphanet <-> HPO/pheocode mapping}
\format{a list of:
  1. table: a data.frame with
     column gene (with Ensembl ID, e.g. ENSG00000084754) ,
     column HPO (HPO term in the format HP:0000512),
     column phecode (e.g. 346),
     column EFO (e.g. EFO:0004207);
  2. script_info: meta information on how data set is generated.}
\source{
\href{https://github.com/hakyimlab/gold-standard/blob/master/scripts/phecode_to_orphanet_to_gene.sh}{Source code generating this silver standard dataset}
}
\usage{
data(orphanet_based_silver_standard)
}
\description{
The Orphanet phenotype to Orphanet gene map and
the Orphanet phenotype to phecode/HPO map
are provided by Lisa Bastarache.
}
\examples{
data(orphanet_based_silver_standard)
head(orphanet_based_silver_standard$table)

}
\keyword{datasets}
