% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silver_standard.R
\name{silver_standard_proto}
\alias{silver_standard_proto}
\title{Plotting PR/ROC curves}
\usage{
silver_standard_proto(
  score_table,
  map_table,
  silver_standard,
  score_cols = NULL,
  trait_code = "phecode",
  mapping_method = "greedy_map"
)
}
\arguments{
\item{score_table}{a data.frame containing 'trait' and 'gene' as columns along with
multiple columns for scores calculated by different method. Note that we interpret HIGHER score as
more likely to be causal. If you have the opposite, please flip it.}

\item{map_table}{a data.frame containing 'trait' as column along with 'phecode' and/or 'HPO' as columns.
This table serves as the map between trait name in 'trait' column and trait codes ('phecode', 'HPO', etc).
So that it is used to map trait names (in score_table) to trait code (in silver_standard).}

\item{silver_standard}{shared along with the package. It is a list with meta information and
a table (a data.frame) where each row pairs trait in phecode, HPO to gene.
New silver_standard data set is awaiting for contribution from the community.}

\item{score_cols}{a vector of column names in score_table to be used as score columns. If it is NULL, all
but 'trait' and 'gene' columns will be used.}

\item{trait_code}{trait code to use for mapping: 'phecode' or 'HPO'.}

\item{mapping_method}{the name of the mapping function used to match trait name with gene
on the basis of map_table and silver_standard$table
New mapping_method is awaiting for contribution from the community.}
}
\value{
list containing PR and ROC plots which are ggplot2 objects
}
\description{
Given a silver standard as 'truth', plot PR/ROC curves for the trait/gene pairs and scores listed in score_table, e.g.
}
\details{
> score_table = data.frame(trait = ..., gene = ..., score1 = ..., score2 = ..., ...)

Silver standard supports HPO, phecode.
User should provide the corresponding HPO or phecode for their GWAS traits in map_table, e.g.

> map_table1 = data.frame(trait = ..., phecode = ...)

> map_table2 = data.frame(trait = ..., HPO = ...)
}
\examples{
silver_standard_proto(
  score_table = data.frame(
    trait = c(rep('t1', 20), rep('t2', 30), rep('t3', 50)),
    gene = paste0('g', sample(1:100, size = 100, replace = TRUE)),
    score1 = runif(100),
    score2 = runif(100),
    stringsAsFactors = FALSE
  ),
  map_table = data.frame(
    trait = c('t1', 't2', 't3'),
    phecode = c('1.1', '23', '12'),
    stringsAsFactors = FALSE
  ),
  silver_standard = list(
    table = data.frame(
      phecode = c(rep('1.1', 10), rep('23', 10), rep('12', 10), rep('2', 10)),
      gene = paste0('g', sample(1:100, size = 40, replace = TRUE)),
      stringsAsFactors = FALSE
    ),
    script_info = 'toy_example'
  )
)

}
